+++ pkg/routing/memory.go
@@ -5,6 +5,7 @@ import (
 	"net/netip"
 	"slices"
 	"sync"
+	"sync/atomic"
 )
 
 var _ Router = &MemoryRouter{}
@@ -12,21 +13,25 @@ var _ Router = &MemoryRouter{}
 type MemoryRouter struct {
 	resolver map[string][]netip.AddrPort
 	self     netip.AddrPort
+	ready    atomic.Bool
 	mx       sync.RWMutex
 }
 
 func NewMemoryRouter(resolver map[string][]netip.AddrPort, self netip.AddrPort) *MemoryRouter {
-	return &MemoryRouter{
+	r := &MemoryRouter{
 		resolver: resolver,
 		self:     self,
 	}
+	r.ready.Store(true)
+	return r
+}
+
+func (m *MemoryRouter) SetReadiness(ready bool) {
+	m.ready.Store(ready)
 }
 
 func (m *MemoryRouter) Ready(ctx context.Context) (bool, error) {
-	m.mx.RLock()
-	defer m.mx.RUnlock()
-
-	return len(m.resolver) > 0, nil
+	return m.ready.Load(), nil
 }
 
 func (m *MemoryRouter) Lookup(ctx context.Context, key string, count int) (Balancer, error) {